+++ src/business/service/DeleteCommentService.java
@@ -2,6 +2,7 @@ package business.service;
 
 import java.io.PrintWriter;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
@@ -12,7 +13,7 @@ import net.sf.json.JSONObject;
 import business.dao.CommentDao;
 import business.dao.ReviewDao;
 import business.dao.SalonDao;
-
+import common._model.TReviewInfo;
 import common.constant.Constant;
 import common.util.DBConnection;
 
@@ -38,27 +39,55 @@ public class DeleteCommentService implements IServiceExcuter{
 			JSONObject jsonObject = new JSONObject();
 
 			int reviewId = -1;
-			if(conn!=null){
+			if(conn!=null && t_commentId!=null){
 				ReviewDao reviewDao = new ReviewDao();
 				CommentDao commentDao = new CommentDao();
-				reviewId = commentDao.getReviewId(dbConnection,t_commentId);
+				//reviewId = commentDao.getReviewId(dbConnection,t_commentId);
+				reviewId = commentDao.get(dbConnection,Integer.parseInt(t_commentId)).getTCommentReviewId();
 				/*
+				 * TODO コメントを削除できる人をレビュアーだけにするか？
 				int reviewUserId = reviewDao.getReviewUserId(dbConnection, t_reviewId);
 				if(reviewUserId == userId){
 					result = true;
 				}
 				*/
+				/**
+				 * レビューカラムからコメントidを削除
+				 */
+				TReviewInfo reviewInfo = new TReviewInfo();
+				reviewInfo = reviewDao.get(dbConnection, reviewId);
+				String commentIds = reviewInfo.getTReviewCommentId();
+				List<String> commentIdList = Arrays.asList(commentIds.split(","));
+				commentIds = "";
+				for(int index=0;index<commentIdList.size();index++){
+					if(!commentIdList.get(index).equals(t_commentId)){
+						commentIds += commentIdList.get(index)+",";
+					}
+				}
+				commentIds = commentIds.substring(0,commentIds.length()-1);
+				reviewInfo.setTReviewCommentId(commentIds);
+				int resultInt = reviewDao.update(dbConnection, reviewInfo);
+				/**
+				 * コメントを削除
+				 */
+				if(resultInt>0){
+					resultInt = commentDao.logicalDelete(dbConnection, Integer.parseInt(t_commentId));
+					if(resultInt>0) result=true;
+				}
+				/*
 				result = reviewDao.DeleteCommentId(
 						dbConnection,
 						reviewId,
 						t_commentId
 						);
+				
 				if(result){
 					result = commentDao.DeleteCommentId(
 							dbConnection,
 							t_commentId
 							);
 				}
+				*/
 				dbConnection.close();
 			}else{
 				responseStatus = HttpServletResponse.SC_INTERNAL_SERVER_ERROR;
