+++ src/business/service/DeleteMenuInfoService.java
@@ -1,6 +1,9 @@
 package business.service;
 
 import java.io.PrintWriter;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -8,6 +11,10 @@ import javax.servlet.http.HttpSession;
 
 import net.sf.json.JSONObject;
 import business.dao.MenuDao;
+import business.dao.SalonDao;
+import business.dao.StylistDao;
+import common._model.THairSalonMasterInfo;
+import common._model.TStylistInfo;
 import common.constant.Constant;
 import common.util.DBConnection;
 
@@ -17,13 +24,6 @@ public class DeleteMenuInfoService implements IServiceExcuter{
 			HttpServletResponse response){
 		
         int responseStatus = HttpServletResponse.SC_OK;
-        /*
-        int userId = request.getHeader(Constant.HEADER_USERID)!= null 
-        		?Integer.parseInt(request.getHeader(Constant.HEADER_USERID)) : -1;
-		 // userIdがパラメータ。なかったら-1を入れておく。
-        //TODO テスト用
-        userId = 1;
-        */
         
   		HttpSession session = request.getSession(false);
 		//salonId kokokara
@@ -55,13 +55,68 @@ public class DeleteMenuInfoService implements IServiceExcuter{
 			boolean result = false;
 			JSONObject jsonObject = new JSONObject();
 			
-			if(conn!=null){
+			if(conn!=null && t_menu_menuId!=null){
 				MenuDao menuDao = new MenuDao();
+				SalonDao salonDao = new SalonDao();
+				StylistDao stylistDao = new StylistDao();
+				
+				/*
 				result = menuDao.DeleteMenuInfoForMaster(
 						dbConnection,
 						t_menu_menuId,
 						salonId
 						);
+						*/
+				int resultInt = menuDao.logicalDelete(dbConnection, Integer.parseInt(t_menu_menuId));
+				if(resultInt>0){ 
+					/**
+					 * SalonからmenuIdを削除
+					 */
+					THairSalonMasterInfo salonInfo = new THairSalonMasterInfo();
+					salonInfo = salonDao.get(dbConnection, salonId);
+					String menuIds = salonInfo.getTHairSalonMasterMenuId();
+					List<String> menuIdList = Arrays.asList(menuIds.split(","));
+					menuIds = "";
+					for(int index=0;index<menuIdList.size();index++){
+						if(!menuIdList.get(index).equals(t_menu_menuId)){
+							menuIds += menuIdList.get(index)+",";
+						}
+					}
+					menuIds = menuIds.substring(0,menuIds.length()-1);
+					salonInfo.setTHairSalonMasterMenuId(menuIds);
+					resultInt = salonDao.update(dbConnection, salonInfo);
+				}
+				if(resultInt>0){ 
+					/**
+					 * StylistカラムからmenuIdを削除
+					 */
+					List<String> stylistIdList = 
+							Arrays.asList(salonDao.get(dbConnection, salonId).getTHairSalonMasterStylistId().split(","));
+					List<TStylistInfo> sList = new ArrayList<TStylistInfo>();
+					for(String id : stylistIdList){
+						TStylistInfo stylistInfo = new TStylistInfo();
+						stylistInfo = stylistDao.get(dbConnection, Integer.parseInt(id));
+						List<String> stylistMenuIdList = Arrays.asList(stylistInfo.getTStylistMenuId().split(","));
+						if(stylistMenuIdList.contains(id)){
+							sList.add(stylistInfo);
+						}
+					}
+					for(TStylistInfo stylistInfo : sList){
+						String menuIds = stylistInfo.getTStylistMenuId();
+						List<String> menuIdList = Arrays.asList(menuIds.split(","));
+						menuIds = "";
+						for(int index=0;index<menuIdList.size();index++){
+							if(!menuIdList.get(index).equals(t_menu_menuId)){
+								menuIds += menuIdList.get(index)+",";
+							}
+						}
+						menuIds = menuIds.substring(0,menuIds.length()-1);
+						stylistInfo.setTStylistMenuId(menuIds);
+						resultInt = stylistDao.update(dbConnection, stylistInfo);
+						if(resultInt<=0) break;
+					}
+					if(resultInt>0) result = true;					
+				}
 				dbConnection.close();
 			}else{
 				responseStatus = HttpServletResponse.SC_INTERNAL_SERVER_ERROR;
